\name{COSMIC.build57-package}
\alias{COSMIC.build57-package}
\alias{COSMIC.build57}
\docType{package}
\title{
Container for COSMIC Build 57.
}
\description{
Container for COSMIC Build 57.
}
\details{
\tabular{ll}{
Package: \tab COSMIC.build57\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-05\cr
}
}
\author{
Sean Davis <seandavi@gmail.com>
}
\references{
  \url{http://ftp://ftp.sanger.ac.uk/pub/CGP/cosmic/data_export/}
  
  "The mutation data was obtained from the Sanger Institute Catalogue Of Somatic 
  Mutations In Cancer web site, http://www.sanger.ac.uk/cosmic 
  Bamford et al (2004) The COSMIC (Catalogue of Somatic Mutations in Cancer) database 
  and website. Br J Cancer, 91,355-358."
}
\keyword{ package }
\seealso{
\code{\linkS4class{GRanges}}
}
\examples{
data(cosmicGRCh37)
cosmicGRCh37
data(cosmicNCBI36)
cosmicNCBI36
}
