\name{cosmicNCBI36}
\alias{cosmicNCBI36}
\docType{data}
\title{
COSMIC data mapped to NCBI36 (hg18)
}
\description{
COSMIC data mapped to NCBI36 (hg18)
}
\usage{data(cosmicNCBI36)}
\format{
  The format is:
Formal class 'GRanges' [package "GenomicRanges"] with 6 slots
  ..@ seqnames       :Formal class 'Rle' [package "IRanges"] with 4 slots
  .. .. ..@ values         : Factor w/ 24 levels "chr1","chr10",..: 2 4 15 18 4 12 21 19 22 9 ...
  .. .. ..@ lengths        : int [1:8393] 1 6 1 1 3 1 1 4 1 5 ...
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ ranges         :Formal class 'IRanges' [package "IRanges"] with 6 slots
  .. .. ..@ start          : int [1:34281] 52245937 9150431 9150104 9116623 9121242 8912192 8882264 41418956 139332477 51987696 ...
  .. .. ..@ width          : int [1:34281] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..@ NAMES          : NULL
  .. .. ..@ elementType    : chr "integer"
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ strand         :Formal class 'Rle' [package "IRanges"] with 4 slots
  .. .. ..@ values         : Factor w/ 3 levels "+","-","*": 3
  .. .. ..@ lengths        : int 34281
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ elementMetadata:Formal class 'DataFrame' [package "IRanges"] with 6 slots
  .. .. ..@ rownames       : NULL
  .. .. ..@ nrows          : int 34281
  .. .. ..@ listData       :List of 24
  .. .. .. ..$ Gene.name                      : Factor w/ 15198 levels "A1CF","A2M","A2ML1",..: 1 2 2 2 2 3 3 4 5 6 ...
  .. .. .. ..$ Accession.Number               : Factor w/ 15158 levels "AB023190","AB051467.1",..: 9114 10964 10964 10964 10964 4598 4598 1096 801 261 ...
  .. .. .. ..$ HGNC.ID                        : int [1:34281] 24086 7 7 7 7 23336 23336 18149 17968 13666 ...
  .. .. .. ..$ Sample.name                    : Factor w/ 153739 levels " HOG06-122"," HOG06-1221",..: 118608 118525 148952 148957 148996 149344 149368 118524 118566 149234 ...
  .. .. .. ..$ ID_sample                      : int [1:34281] 1312984 1312971 1287214 1287216 1287244 1475013 1474986 1312970 1312975 1474963 ...
  .. .. .. ..$ ID_tumour                      : int [1:34281] 1223893 1223880 1198492 1198494 1198522 1398712 1398685 1223879 1223884 1398662 ...
  .. .. .. ..$ Primary.site                   : Factor w/ 43 levels "Fallopian tube",..: 8 8 8 8 8 21 21 8 8 21 ...
  .. .. .. ..$ Site.subtype                   : Factor w/ 172 levels "NS","abdomen",..: 16 16 16 16 16 1 1 16 16 1 ...
  .. .. .. ..$ Primary.histology              : Factor w/ 84 levels "Ewings_sarcoma-peripheral_primitive_neuroectodermal_tumour",..: 40 40 40 40 40 23 23 40 40 23 ...
  .. .. .. ..$ Histology.subtype              : Factor w/ 399 levels "ACTH","ACTH-FSH",..: 82 82 82 82 82 351 351 33 82 351 ...
  .. .. .. ..$ Genome.wide.screen             : Factor w/ 3 levels "","n","y": 1 1 1 1 1 3 3 1 1 3 ...
  .. .. .. ..$ Mutation.ID                    : int [1:34281] 40879 39653 35965 35963 35964 78858 79446 39410 41045 79447 ...
  .. .. .. ..$ Mutation.CDS                   : Factor w/ 38881 levels "","Intronic",..: 38446 37309 1522 22775 20223 23591 37411 37952 19219 6125 ...
  .. .. .. ..$ Mutation.AA                    : Factor w/ 54163 levels "","p.","p.(1409)ins6",..: 9502 47270 15450 11774 46083 36292 1714 50231 215 44522 ...
  .. .. .. ..$ Mutation.Description           : Factor w/ 15 levels "Complex - compound substitution",..: 11 11 11 11 13 12 13 11 13 13 ...
  .. .. .. ..$ Mutation.zygosity              : Factor w/ 3 levels "het","hom","unk": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ Mutation.NCBI36.genome.position: Factor w/ 34282 levels "","10:100001319-100001319",..: 668 5102 5101 5098 5099 5086 5085 18559 23172 4666 ...
  .. .. .. ..$ Mutation.NCBI36.strand         : Factor w/ 3 levels "","+","-": 3 3 3 3 3 2 2 3 3 3 ...
  .. .. .. ..$ Mutation.GRCh37.strand         : Factor w/ 3 levels "","+","-": 3 3 3 3 3 2 2 3 3 3 ...
  .. .. .. ..$ Mutation.somatic.status        : Factor w/ 6 levels "Confirmed germline variant",..: 2 2 2 2 2 2 2 2 2 2 ...
  .. .. .. ..$ Pubmed_PMID                    : int [1:34281] 18772396 18772396 18772890 18772890 18772890 21720365 21720365 18772396 18772396 21720365 ...
  .. .. .. ..$ Sample.source                  : Factor w/ 27 levels "NS","autopsy - NOS",..: 23 23 24 24 24 1 1 27 27 1 ...
  .. .. .. ..$ Tumour.origin                  : Factor w/ 8 levels "NS","adenoma adjacent to primary tumour",..: 7 5 6 7 7 5 5 5 5 5 ...
  .. .. .. ..$ Comments                       : Factor w/ 2723 levels "","Age at tumour recurrence:1 year after diagnosis of TMD",..: 1002 1002 1 1 1 1 1 1002 1002 1 ...
  .. .. ..@ elementType    : chr "ANY"
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ seqinfo        :Formal class 'Seqinfo' [package "GenomicRanges"] with 4 slots
  .. .. ..@ seqnames   : chr [1:24] "chr1" "chr10" "chr11" "chr12" ...
  .. .. ..@ seqlengths : int [1:24] NA NA NA NA NA NA NA NA NA NA ...
  .. .. ..@ is_circular: logi [1:24] NA NA NA NA NA NA ...
  .. .. ..@ genome     : chr [1:24] NA NA NA NA ...
  ..@ metadata       : list()
}
\details{
The data are in a GRanges object with elementMetadata from the original
download.  Note that some of the ranges are large, so one may need to
filter when looking for overlapping variants.
}
\source{
\url{ftp://ftp.sanger.ac.uk/pub/CGP/cosmic/data_export/CosmicMutantExport_v57_180112.tsv}
}
\references{
\url{http://www.sanger.ac.uk/genetics/CGP/cosmic/}
}
\examples{
data(cosmicGRCh37)
}
\keyword{datasets}
